Fix tests by moving computeFilePath from DiskManager to fsUtils
