+++ creator-node/src/diskManager.ts
@@ -2,7 +2,6 @@ import type Logger from 'bunyan'
 
 import path from 'path'
 import fs from 'fs-extra'
-import CID from 'cids'
 import { chunk, range } from 'lodash'
 
 import DbManager from './dbManager'
@@ -14,7 +13,11 @@ import {
   execShellCommand,
   runShellCommand,
   verifyCIDMatchesExpected,
-  timeout
+  timeout,
+  ensureDirPathExists,
+  computeFilePath,
+  computeFilePathAndEnsureItExists,
+  computeFilePathInDirAndEnsureItExists
 } from './utils'
 
 const models = require('./models')
@@ -82,160 +85,6 @@ export async function getTmpTrackUploadArtifactsPath() {
     TMP_TRACK_ARTIFACTS_CREATED = true
   }
   return dirPath
-}
-
-function _getStorageLocationForCID(cid: string) {
-  try {
-    CID.isCID(new CID(cid))
-  } catch (e: any) {
-    tracing.recordException(e)
-    genericLogger.error(`CID invalid, cid=${cid}, error=${e.toString()}`)
-    throw new Error(`Please pass in a valid cid. Passed in ${cid} ${e.message}`)
-  }
-
-  // This is the directory path that file with cid will go into.
-  // The reason for nesting `files` inside `/file_storage` is because legacy nodes store files at the root of `/file_storage`, and
-  // that can cause potential collisions if we're creating large amounts of subdirectories. A way to mitigate this is create one
-  // directory in the root `/file_storage` and all other directories inside of it like `file_storage/files/<directoryID>/<cid>
-  const directoryID = cid.slice(-4, -1)
-  const storageLocationForCid = path.join(
-    getConfigStoragePath(),
-    'files',
-    directoryID
-  )
-  // in order to easily dev against the older and newer paths, the line below is the legacy storage path
-  // const storageLocationForCid = getConfigStoragePath()
-
-  return storageLocationForCid
-}
-
-/**
- * Construct the path to a file or directory given a CID.
- * This function does not ensure the path returned exists on disk, so use it for read-only operations
- * (like /ipfs retrieval) or tasks where a subsequent step will already ensure the path exists (like syncs).
- *
- * eg. if you have a file CID `Qmabcxyz`, use this function to get the path /file_storage/files/cxy/Qmabcxyz
- * eg. if you have a dir CID `Qmdir123`, use this function to get the path /file_storage/files/r12/Qmdir123/
- * Use `computeFilePathInDir` if you want to get the path for a file inside a directory.
- *
- * @dev Returns a path with the three characters before the last character
- *      eg QmYfSQCgCwhxwYcdEwCkFJHicDe6rzCAb7AtLz3GrHmuU6 will be eg /file_storage/muU/QmYfSQCgCwhxwYcdEwCkFJHicDe6rzCAb7AtLz3GrHmuU6
- * @param {String} cid file system destination, either filename or directory
- */
-export function computeFilePath(cid: string) {
-  const storageLocationForCid = _getStorageLocationForCID(cid)
-  return path.join(storageLocationForCid, cid)
-}
-
-/**
- * Use for operations where it's necessary to ensure that the path exists before using it.
- * @see computeFilePath - does the same thing but also performs the equivalent of 'mkdir -p'
- */
-export async function computeFilePathAndEnsureItExists(cid: string) {
-  const parentDirPath = _getStorageLocationForCID(cid)
-
-  // create the subdirectories in parentDirHash if they don't exist
-  await ensureDirPathExists(parentDirPath)
-
-  return path.join(parentDirPath, cid)
-}
-
-/**
- * Construct the legacy path to a file or directory given a CID
- */
-export function computeLegacyFilePath(cid: string) {
-  if (!isValidCID(cid)) {
-    throw new Error(`[computeLegacyFilePath] [CID=${cid}] Invalid CID.`)
-  }
-  return path.join(getConfigStoragePath(), cid)
-}
-
-/**
- * Boolean function to check if arg is a valid CID
- */
-export function isValidCID(cid: string) {
-  try {
-    // Will throw if `new CID(cid)` fails
-    // CID.isCID() returns boolean
-    return CID.isCID(new CID(cid))
-  } catch (e) {
-    return false
-  }
-}
-
-function _validateFileAndDir(dirName: string, fileName: string) {
-  if (!dirName || !fileName) {
-    genericLogger.error(
-      `Invalid dirName and/or fileName, dirName=${dirName}, fileName=${fileName}`
-    )
-    throw new Error('Must pass in valid dirName and fileName')
-  }
-
-  try {
-    CID.isCID(new CID(dirName))
-    CID.isCID(new CID(fileName))
-  } catch (e: any) {
-    genericLogger.error(
-      `CID invalid, dirName=${dirName}, fileName=${fileName}, error=${e.toString()}`
-    )
-    throw new Error(
-      `Please pass in a valid cid for dirName and fileName. Passed in dirName: ${dirName} fileName: ${fileName} ${e.message}`
-    )
-  }
-}
-
-/**
- * Given a directory name and a file name, construct the full file system path for a directory and a folder inside a directory.
- * This function does not ensure the path returned exists on disk, so use it for read-only operations
- * (like /ipfs retrieval) or tasks where a subsequent step will already ensure the path exists (like syncs).
- *
- * eg if you're manually computing the file path to an file `Qmabcxyz` inside a dir `Qmdir123`, use this function to get the
- * path with both the dir and the file /file_storage/files/r12/Qmdir123/Qmabcxyz
- * Use `computeFilePath` if you just want to get to the path of a file or directory.
- *
- * @param {String} dirName directory name
- * @param {String} fileName file name
- */
-export function computeFilePathInDir(dirName: string, fileName: string) {
-  _validateFileAndDir(dirName, fileName)
-
-  const parentDirPath = computeFilePath(dirName)
-  const absolutePath = path.join(parentDirPath, fileName)
-  genericLogger.info(`File path computed, absolutePath=${absolutePath}`)
-  return absolutePath
-}
-
-/**
- * Use for operations where it's necessary to ensure that the path exists before using it.
- * @see computeFilePathInDir - does the same thing but also performs the equivalent of 'mkdir -p'
- */
-export async function computeFilePathInDirAndEnsureItExists(
-  dirName: string,
-  fileName: string
-) {
-  _validateFileAndDir(dirName, fileName)
-
-  const parentDirPath = await computeFilePathAndEnsureItExists(dirName)
-  const absolutePath = path.join(parentDirPath, fileName)
-  genericLogger.info(`File path computed, absolutePath=${absolutePath}`)
-  return absolutePath
-}
-
-/**
- * Given a directory path, this function will create the dirPath if it doesn't exist
- * If it does exist, it will not overwrite, effectively a no-op
- * @param {*} dirPath fs directory path to create if it does not exist
- */
-export async function ensureDirPathExists(dirPath: string) {
-  try {
-    // the mkdir recursive option is equivalent to `mkdir -p` and should created nested folders several levels deep
-    await fs.mkdir(dirPath, { recursive: true })
-  } catch (e: any) {
-    genericLogger.error(
-      `Error making directory, dirName=${dirPath}, error=${e.toString()}`
-    )
-    throw new Error(`Error making directory at ${dirPath} - ${e.message}`)
-  }
 }
 
 /**
