+++ creator-node/test/dbManager.test.js
@@ -733,7 +733,7 @@ describe('Test deleteAllCNodeUserDataFromDB()', async function () {
           '../../fileManager': {
             copyMultihashToFs: sinon
               .stub(FileManager, 'copyMultihashToFs')
-              .returns(await DiskManager.computeFilePathAndEnsureItExists(mockCid)),
+              .returns(await utils.computeFilePathAndEnsureItExists(mockCid)),
             '@global': true
           }
         }
@@ -1202,7 +1202,7 @@ describe('Test fixInconsistentUser()', async function () {
         sourceFile: null,
         fileName: null,
         dirMultihash: null,
-        storagePath: await DiskManager.computeFilePath(metadataCID),
+        storagePath: await utils.computeFilePath(metadataCID),
         type: "metadata",
         clock: 1,
         skipped: false
@@ -1250,7 +1250,7 @@ describe('Test fixInconsistentUser()', async function () {
         sourceFile: null,
         fileName: null,
         dirMultihash: null,
-        storagePath: await DiskManager.computeFilePath(metadataCID),
+        storagePath: await utils.computeFilePath(metadataCID),
         type: "metadata",
         clock: 1,
         skipped: false
@@ -1317,7 +1317,7 @@ describe('Test fixInconsistentUser()', async function () {
         sourceFile: null,
         fileName: null,
         dirMultihash: null,
-        storagePath: await DiskManager.computeFilePath(metadataCID),
+        storagePath: await utils.computeFilePath(metadataCID),
         type: "metadata",
         clock: 1,
         skipped: false
@@ -1365,7 +1365,7 @@ describe('Test fixInconsistentUser()', async function () {
         sourceFile: null,
         fileName: null,
         dirMultihash: null,
-        storagePath: await DiskManager.computeFilePath(metadataCID),
+        storagePath: await utils.computeFilePath(metadataCID),
         type: "metadata",
         clock: 1,
         skipped: false
