+++ creator-node/test/fileManager.test.js
@@ -20,7 +20,10 @@ const { sortKeys } = require('../src/apiSigning')
 const DiskManager = require('../src/diskManager')
 const { getLibsMock } = require('./lib/libsMock')
 const DecisionTree = require('../src/utils/decisionTree')
-const { deleteAttemptedStateFixes } =require('../src/utils')
+const {
+  deleteAttemptedStateFixes,
+  computeFilePathAndEnsureItExists
+} = require('../src/utils')
 
 const storagePath = config.get('storagePath')
 
@@ -548,7 +551,7 @@ describe('test fileManager', () => {
 
       // 1 segment should be saved in <storagePath>/QmSMQGu2vrE6UwXiZDCxyJwTsCcpPrYNBPJBL4by4LKukd
       const segmentCID = 'QmSMQGu2vrE6UwXiZDCxyJwTsCcpPrYNBPJBL4by4LKukd'
-      const syncedSegmentPath = await DiskManager.computeFilePathAndEnsureItExists(segmentCID)
+      const syncedSegmentPath = await computeFilePathAndEnsureItExists(segmentCID)
       assert.ok(await fs.pathExists(syncedSegmentPath))
 
       // the segment content should match the original sourcefile
@@ -644,7 +647,7 @@ describe('test fileManager', () => {
       }
 
       // check that the metadata file was written to storagePath under its multihash
-      const metadataPath = await DiskManager.computeFilePathAndEnsureItExists(resp.cid)
+      const metadataPath = await computeFilePathAndEnsureItExists(resp.cid)
       assert.ok(await fs.pathExists(metadataPath))
 
       // check that the contents of the metadata file is what we expect
