+++ creator-node/test/playlists.test.js
@@ -11,6 +11,7 @@ const DiskManager = require('../src/diskManager')
 const { createStarterCNodeUser } = require('./lib/dataSeeds')
 const { getLibsMock } = require('./lib/libsMock')
 const { sortKeys } = require('../src/apiSigning')
+const { computeFilePathAndEnsureItExists } = require('../src/utils')
 
 describe('Test Playlists', function () {
   let app, server, session, libsMock
@@ -80,7 +81,7 @@ describe('Test Playlists', function () {
     }
 
     // check that the metadata file was written to storagePath under its multihash
-    const metadataPath = await DiskManager.computeFilePathAndEnsureItExists(
+    const metadataPath = await computeFilePathAndEnsureItExists(
       resp.body.data.metadataMultihash
     )
     assert.ok(await fs.pathExists(metadataPath))
@@ -112,7 +113,7 @@ describe('Test Playlists', function () {
       .expect(200)
 
     // check that the metadata file was written to storagePath under its multihash
-    const metadataPath = await DiskManager.computeFilePathAndEnsureItExists(
+    const metadataPath = await computeFilePathAndEnsureItExists(
       resp.body.data.metadataMultihash
     )
     assert.ok(await fs.pathExists(metadataPath))
@@ -157,7 +158,7 @@ describe('Test Playlists', function () {
       .set('User-Id', session.userId)
 
     const imageDirCID = resp.body.data.dirCID
-    const playlistImagePath = await DiskManager.computeFilePathAndEnsureItExists(imageDirCID)
+    const playlistImagePath = await computeFilePathAndEnsureItExists(imageDirCID)
     assert.ok(await fs.pathExists(playlistImagePath))
     const fileRecord = await models.File.findOne({
       where: {
@@ -182,7 +183,7 @@ describe('Test Playlists', function () {
       .expect(200)
 
     // check that the metadata file was written to storagePath under its multihash
-    const metadataPath = await DiskManager.computeFilePathAndEnsureItExists(
+    const metadataPath = await computeFilePathAndEnsureItExists(
       resp2.body.data.metadataMultihash
     )
     assert.ok(await fs.pathExists(metadataPath))
